+++ src/monaca.js
@@ -19,6 +19,8 @@ colors.setTheme({
   data: 'grey',
   help: 'cyan',
   warn: ['yellow', 'bold'],
+  commands: ['cyan', 'italic'],
+  url: ['cyan', 'underline'],
   debug: 'blue',
   error: ['red', 'bold'],
   success: ['green', 'bold']