+++ src/remote.js
@@ -2,7 +2,7 @@
 'use strict';
 
 var path = require('path'),
-  open = require('open'),
+  open = require('opn'),
   argv = require('optimist').argv,
   shell = require('shelljs'),
   Monaca = require('monaca-lib').Monaca,
@@ -80,9 +80,11 @@ RemoteTask.remote = function(task) {
       }
     )
     // Assuring this is a Monaca-like project.
-    .then(
-      function(info) {
+    .then( 
+      (info) => {
         projectInfo = info;
+        lib.needToUpgrade(cwd, monaca);
+
         error = 'Upload failed: ';
         if (!params['build-list']) {
           return monaca.uploadProject(cwd)
@@ -109,15 +111,18 @@ RemoteTask.remote = function(task) {
             .then(
               function(url) {
                 var deferred = Q.defer();
-                open(url, function(error) {
-                  if (error) {
-                    return deferred.reject(error);
-                  }
+                open(url, {wait: false})
+                .then(() => {
                   if (task === 'config') {
                     util.warn('\nOnce the Cloud configuration has been saved, run `monaca download` to get the changes locally.');
                   }
                   deferred.resolve();
-                });
+                })
+                .catch(error => {
+                  if (error) {
+                    return deferred.reject(error);
+                  }
+                }) ;
                 return deferred.promise;
               }
             )