+++ src/transpile.js
@@ -1,73 +1,43 @@
-var path = require('path'),
-  fs = require('fs'),
-  exec = require('child_process').exec,
-  util = require(path.join(__dirname, 'util')),
-  lib = require(path.join(__dirname, 'lib')),
-  Monaca = require('monaca-lib').Monaca;
+const path = require('path');
+const util = require(path.join(__dirname, 'util'));
+const lib = require(path.join(__dirname, 'lib'));
+const Monaca = require('monaca-lib').Monaca;
 
-var TranspileTask = {}, monaca;
+let TranspileTask = {}; let monaca;
 
 TranspileTask.run = function(taskName, info) {
+  let projectDir;
   monaca = new Monaca(info);
 
   lib.findProjectDir(process.cwd(), monaca)
+  // Checking if the user needs to upgrade the project
   .then(
-    function(dir) {
-      var projectDir = dir,
-        projectConfig;
+    (dir) => {
+      projectDir= dir;
+      lib.needToUpgrade(projectDir, monaca);
 
-      try {
-        projectConfig = require(path.join(projectDir, 'package.json'));
-      } catch (err) {
-        return transpileProject(taskName, projectDir);
-      }
+      if (!monaca.hasTranspileScript(projectDir)) util.fail('This project is not transpilable.');
+      else util.checkNodeRequirement();
 
-      if (projectConfig && projectConfig.scripts && projectConfig.scripts.build) {
-        var childProcess = exec('npm run build');
+      let report = {
+        event: 'transpile'
+      };
+    
+      monaca.reportAnalytics(report);
 
-        childProcess.stdout.on('data', function(data) {
-          process.stdout.write(data.toString());
-        });
+    return monaca.transpile(projectDir)
+      .then(
+        monaca.reportFinish.bind(monaca, report),
+        monaca.reportFail.bind(monaca, report)
+      )
+      .then(
+        util.success.bind(null),
+        util.fail.bind(null, 'Project has failed to transpile. ')
+      );
 
-        childProcess.stderr.on('data', function(data) {
-          if (data) {
-            process.stderr.write(data.toString().error);
-          }
-        });
-      } else {
-        return transpileProject(taskName, projectDir);
-      }
     }
   )
   .catch(util.fail.bind(null, 'Project ' + taskName + ' failed: '));
-};
-
-var transpileProject = function(taskName, projectDir) {
-  if (!monaca.isTranspilable(projectDir)) {
-    util.fail('This project is not transpilable.');
-  }
-
-  var report = {
-    event: 'transpile'
-  };
-
-  monaca.reportAnalytics(report);
-
-  var isTranspileEnabled = monaca.isTranspileEnabled(projectDir);
-
-  if (isTranspileEnabled) {
-    util.checkNodeRequirement();
-  }
-
-  return monaca.transpile(projectDir)
-    .then(
-      monaca.reportFinish.bind(monaca, report),
-      monaca.reportFail.bind(monaca, report)
-    )
-    .then(
-      util.success.bind(null),
-      util.fail.bind(null, 'Project has failed to transpile. ')
-    );
 };
 
 module.exports = TranspileTask;